different IO API: read and parse
